Split out BlurredImage component
