+++ app/pages/home-for-user/blurred-image.styl
@@ -1,0 +1,17 @@
+.blurred-image
+  &__container
+    position: relative
+    min-width: 1em
+    min-height: 1em
+    overflow: hidden
+
+  &__display
+    background-position: 50% 50%
+    background-size: cover
+    bottom: -1em
+    -webkit-filter: blur(1em)
+    filter: blur(1em)
+    left: -1em
+    position: absolute
+    right: -1em
+    top: -1em
