+++ app/pages/home-for-user/index.styl
@@ -1,22 +1,3 @@
-.blurred-image
-  &__container
-    position: relative
-    min-width: 1em
-    min-height: 1em
-    overflow: hidden
-
-  &__display
-    background-position: 50% 50%
-    background-size: cover
-    bottom: -1em
-    -webkit-filter: blur(1em)
-    filter: blur(1em)
-    left: -1em
-    opacity: 0.5
-    position: absolute
-    right: -1em
-    top: -1em
-
 .home-page-for-user
   background: black
   color: white
@@ -25,9 +6,10 @@
   position: relative
 
   &__background
-    position: absolute
     height: 100%
     left: 0
+    opacity: 0.5
+    position: absolute
     top: 0
     width: 100%
